chore: extract cli::args test module to its own file (#76)
src/cli/args.rs had grown to ~994 lines: clap definitions plus a
350-line #[cfg(test)] mod tests block. Split it into src/cli/args/mod.rs
(clap definitions) and src/cli/args/tests.rs (tests, declared via
mod tests;), matching the mod.rs convention already used by every other
directory module in this repo (api, cli, commands, db, embed, query,
sync, update). Purely mechanical: no behavior change, no test renames.
